imaging/src/requirements.txt
pydicom == 2.1.2
numpy == 1.19.5
nibabel == 3.2.1
nilearn == 0.7.0
matplotlib == 3.1.3
dcmstack == 0.8.0
pandas == 1.1.5
pyradiomics == 3.0.1
sagemaker == 2.27.0
boto3 == 1.17.18